Pass LambdaRoleArn param to createMacroStack step
The createMacroStack template has a LambdaAssumeRoleNotSpecified
condition that skips creating BasicLambdaRole if a role is passed as a
parameter.

However, no parameters are passed to the createMacroStack.
"LambdaAssumeRoleNotSpecified" is then always true regardless of if the
SSM Document is called with a Lambda role.

The step then fails if the user doesn't have the permissions to create
IAM Roles.

This change adds the necessary parameters to createMacroStack so it's
executed correctly when given a LambdaRoleArn.